[eslint] Reduce max dependencies to 15
