+++ packages/wallet/app/components/profile/header/Avatar.tsx
